Refactor code and add debug logging to file
